minitest helper addition for testing angelo::base subclasses
